+++ lib/CalDAV/Schedule/ISchedulingObject.php
@@ -8,6 +8,6 @@ namespace Sabre\CalDAV\Schedule;
  * @license http://code.google.com/p/sabredav/wiki/License Modified BSD License
  * @copyright Copyright (C) 2007-2014 fruux GmbH. All rights reserved.
  */
-interface ISchedulingObject extends \Sabre\DAV\IFile {
+interface ISchedulingObject extends \Sabre\CalDAV\ICalendarObject {
 
 }